cleanup and added  store for caching flow
