+++ libs/query/src/lib/aa/funManagement/index.ts
@@ -1,0 +1,2 @@
+export * from './service';
+export * from './store';
