+++ libs/query/src/lib/aa/groups/groups.store.ts
@@ -6,7 +6,6 @@ const initialStore = {
   stakeholdersGroupsMeta: {},
   beneficiariesGroups: [],
   beneficiariesGroupsMeta: {},
-  assignedFundData: {},
 };
 
 type BeneficiariesGroupState = {
@@ -19,9 +18,6 @@ type StakeholdersGroupsState = {
   stakeholdersGroupsMeta: any;
 };
 
-type FundAssignmentState = {
-  assignedFundData: any;
-};
 type BeneficiariesGroupsStateAction = {
   setBeneficiariesGroup: (beneficiariesGroup: any) => void;
   setBeneficiariesGroupMeta: (meta: any) => void;
@@ -32,16 +28,11 @@ type StakeholdersGroupsStateAction = {
   setStakeholdersGroupsMeta: (meta: any) => void;
 };
 
-type FundAssignmentStateAction = {
-  setAssignedFundData: (assignedFundData: any) => void;
-};
 type BeneficiariesGroupsStore = BeneficiariesGroupState &
   BeneficiariesGroupsStateAction;
 
 type StakeholdersGroupsStore = StakeholdersGroupsState &
   StakeholdersGroupsStateAction;
-
-type FundAssignmentStore = FundAssignmentState & FundAssignmentStateAction;
 
 export const useBeneficiariesGroupStore =
   zustandStore<BeneficiariesGroupsStore>(
@@ -71,20 +62,6 @@ export const useStakeholdersGroupsStore = zustandStore<StakeholdersGroupsStore>(
     devtoolsEnabled: true,
     persistOptions: {
       name: 'aaStakeholdersGroupStore',
-      storage: localStore,
-    },
-  },
-);
-
-export const useFundAssignmentStore = zustandStore<FundAssignmentStore>(
-  (set) => ({
-    ...initialStore,
-    setAssignedFundData: (assignedFundData) => set({ assignedFundData }),
-  }),
-  {
-    devtoolsEnabled: true,
-    persistOptions: {
-      name: 'aaFundAssignmentStore',
       storage: localStore,
     },
   },
