+++ libs/query/src/lib/aa/index.ts
@@ -7,3 +7,4 @@ export * from './phases';
 export * from './graph';
 export * from './daily-monitoring';
 export * from './communication';
+export * from './funManagement';
