+++ libs/query/src/lib/projects/projects.service.ts
@@ -599,31 +599,6 @@ export const useProjectBeneficiaries = (payload: GetProjectBeneficiaries) => {
   };
 };
 
-export const useFetchTokenStatsStellar = (payload: any) => {
-  const projectBalance = useProjectAction();
-
-  const { projectUUID, ...restPayload } = payload;
-  const restPayloadString = JSON.stringify(restPayload);
-
-  const query = useQuery({
-    queryKey: ['aa.stellar.getStellarStats', restPayloadString],
-    refetchOnMount: true,
-    refetchOnWindowFocus: true,
-    queryFn: async () => {
-      const mutate = await projectBalance.mutateAsync({
-        uuid: projectUUID as '${string}-${string}-${string}-${string}-${string}',
-        data: {
-          action: 'aa.stellar.getStellarStats',
-          payload: {},
-        },
-      });
-      return mutate;
-    },
-  });
-
-  return query;
-};
-
 export const useProjectBeneficiaryDetail = (payload: any) => {
   const q = useProjectAction<Beneficiary>();
   const { projectUUID, ...restPayload } = payload;
